good state of the refactoring
